first restructure of architecture
