+++ src/declare4py/checker_result.py
