+++ src/declare4py/core/log_generator.py
@@ -1,0 +1,7 @@
+
+class LogGenerator:
+
+    def __init__(self):
+        self.log_length = None
+        self.max_events = None
+        self.min_events = None
