+++ src/declare4py/core/monitoring.py
@@ -1,0 +1,4 @@
+class Monitoring(PMTask):
+
+    def __init__(self):
+        pass
