+++ src/declare4py/core/pm_task.py
@@ -1,0 +1,5 @@
+class PMTask:
+
+    def __init__(self):
+        self.log_analyzer = None
+        self.decl_model = None
