+++ src/declare4py/models/conformance_checking_results.py
@@ -1,0 +1,7 @@
+from declare4py.src.declare4py.models.basic_mp_declare_conformance_checking import BasicMPDeclareConformanceChecking
+
+
+class ConformanceCheckingResults:
+
+    def __init__(self, dict_results: dict):
+        self.dict_results = dict_results
