+++ src/declare4py/models/discovery_results.py
@@ -1,0 +1,7 @@
+from declare4py.src.declare4py.models.basic_declare_discovery import BasicDeclareDiscovery
+
+
+class DiscoveryResults:
+
+    def __init__(self):
+        BasicDeclareDiscovery.__init__(self)
