+++ src/declare4py/models/query_checking_results.py
@@ -1,0 +1,7 @@
+from declare4py.src.declare4py.models.basic_mp_declare_query_checking import BasicMPDeclareQueryChecking
+
+
+class QueryCheckingResults:
+
+    def __init__(self):
+        BasicMPDeclareQueryChecking.__init__(self)
