--- src/declare4py/monitoring.py
@@ -1,1 +1,0 @@
-class Monitoring: None
